Restructure the AGN functions in an AGN class
